Refactored 2 packages
Moved ..

com.example.android.common.view
com.hlidskia.android.preference

..to..

com.ichi2.ui

as I couldn't find a reason they should be in rogue packages. Perhaps we should start to migrate packages/classes to ..

com.ichi2.anki ?
