+++ AnkiDroid/src/main/java/com/ichi2/anki/Preferences.java
@@ -48,7 +48,7 @@ import android.widget.LinearLayout;
 import android.widget.Toast;
 
 import com.afollestad.materialdialogs.MaterialDialog;
-import com.hlidskialf.android.preference.SeekBarPreference;
+import com.ichi2.ui.SeekBarPreference;
 import com.ichi2.anim.ActivityTransitionAnimation;
 import com.ichi2.anki.exception.StorageAccessException;
 import com.ichi2.async.DeckTask;
