+++ AnkiDroid/src/main/java/com/ichi2/anki/Statistics.java
@@ -36,7 +36,7 @@ import android.widget.ProgressBar;
 import android.widget.Spinner;
 import android.widget.TextView;
 
-import com.example.android.common.view.SlidingTabLayout;
+import com.ichi2.ui.SlidingTabLayout;
 import com.ichi2.anim.ActivityTransitionAnimation;
 import com.ichi2.anki.stats.AnkiStatsTaskHandler;
 import com.ichi2.anki.stats.ChartView;
