+++ AnkiDroid/src/main/res/layout/card_template_editor_activity.xml
@@ -8,7 +8,7 @@
     <LinearLayout android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical">
-        <com.example.android.common.view.SlidingTabLayout
+        <com.ichi2.ui.SlidingTabLayout
             android:id="@+id/sliding_tabs"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"/>
